use _time_type
